(#66) AnimaPolygonCollider2dBuilderの機能をAnimaCanvasに移植。AnimaPolygonCollider2dBuilder.csを削除。
