--- Assets/Scripts/AnimaPolygonCollider2dBuilder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12682c0e4cb2c294bba64aba01600b64
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
